requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.4
pydantic-settings>=2.2.1
google-genai
supabase>=2.4.0
redis>=5.0.3
pytest>=8.1.1
httpx>=0.27.0
pymupdf
python-docx
python-multipart
groq
